fix: tweak entries
