+++ packages/plugin-rsc/examples/react-router/cf/entry.rsc.tsx
@@ -7,7 +7,3 @@ export default {
     return fetchServer(request)
   },
 }
-
-if (import.meta.hot) {
-  import.meta.hot.accept()
-}
