+++ packages/plugin-rsc/examples/react-router/cf/entry.ssr.tsx
@@ -1,13 +1,9 @@
-import handler from '../react-router-vite/entry.ssr'
+import { generateHTML } from '../react-router-vite/entry.ssr'
 
 console.log('[debug:cf-ssr-entry]')
 
-// TODO:
-// shouldn't "entry.rsc.tsx" be the main server entry
-// and optionally call "entry.ssr.tsx" only for rendering html?
-
 export default {
   fetch(request: Request, env: any) {
-    return handler(request, (request) => env.RSC.fetch(request))
+    return generateHTML(request, (request) => env.RSC.fetch(request))
   },
 }
