+++ packages/plugin-rsc/examples/react-router/react-router-vite/entry.rsc.single.tsx
@@ -1,0 +1,10 @@
+import { fetchServer } from './entry.rsc'
+
+export default async function handler(request: Request) {
+  // Import the generateHTML function from the client environment
+  const ssr = await import.meta.viteRsc.loadModule<
+    typeof import('./entry.ssr')
+  >('ssr', 'index')
+
+  return ssr.generateHTML(request, fetchServer)
+}
