+++ packages/plugin-rsc/examples/react-router/react-router-vite/entry.rsc.tsx
@@ -32,15 +32,6 @@ export function fetchServer(request: Request) {
   })
 }
 
-export default async function handler(request: Request) {
-  // Import the generateHTML function from the client environment
-  const ssr = await import.meta.viteRsc.loadModule<
-    typeof import('./entry.ssr')
-  >('ssr', 'index')
-
-  return ssr.generateHTML(request, fetchServer)
-}
-
 if (import.meta.hot) {
   import.meta.hot.accept()
 }
