+++ packages/plugin-rsc/examples/react-router/react-router-vite/entry.ssr.tsx
@@ -36,10 +36,3 @@ export async function generateHTML(
     },
   })
 }
-
-export default async function handler(
-  request: Request,
-  fetchServer: (request: Request) => Promise<Response>,
-): Promise<Response> {
-  return generateHTML(request, fetchServer)
-}
